+++ module-apps/application-call/widgets/StateIcons.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
-#include "Icons.hpp"
+#include "StateIcons.hpp"
 
 #include "application-call/data/CallAppStyle.hpp"
 
@@ -22,30 +22,31 @@ namespace gui
         constexpr auto speakerStr    = "app_call_speaker";
         constexpr auto speakerOnStr  = "app_call_speaker_on";
 
-        const Icon<AddContactIconState>::IconMap contactIconMap = {
+        const StateIcon<AddContactIconState>::IconMap contactIconMap = {
             {AddContactIconState::ADD_CONTACT, {crossImg, addContactStr}}};
-        const Icon<SendSmsIconState>::IconMap smsIconMap = {{SendSmsIconState::SEND_SMS, {messageImg, sendSmstStr}}};
-        const Icon<MicrophoneIconState>::IconMap microphoneIconMap = {
+        const StateIcon<SendSmsIconState>::IconMap smsIconMap = {
+            {SendSmsIconState::SEND_SMS, {messageImg, sendSmstStr}}};
+        const StateIcon<MicrophoneIconState>::IconMap microphoneIconMap = {
             {MicrophoneIconState::MUTE, {muteImg, muteStr}}, {MicrophoneIconState::MUTED, {mutedImg, mutedStr}}};
-        const Icon<SpeakerIconState>::IconMap speakerIconMap = {
+        const StateIcon<SpeakerIconState>::IconMap speakerIconMap = {
             {SpeakerIconState::SPEAKER, {speakerImg, speakerStr}},
             {SpeakerIconState::SPEAKERON, {speakerOnImg, speakerOnStr}}};
     } // namespace
 
     AddContactIcon::AddContactIcon(Item *parent, std::uint32_t x, std::uint32_t y)
-        : Icon(parent, x, y, callAppStyle::icon::x_margin, AddContactIconState::ADD_CONTACT, contactIconMap)
+        : StateIcon(parent, x, y, callAppStyle::icon::x_margin, AddContactIconState::ADD_CONTACT, contactIconMap)
     {}
 
     SendSmsIcon::SendSmsIcon(Item *parent, std::uint32_t x, std::uint32_t y)
-        : Icon(parent, x, y, callAppStyle::icon::x_margin, SendSmsIconState::SEND_SMS, smsIconMap)
+        : StateIcon(parent, x, y, callAppStyle::icon::x_margin, SendSmsIconState::SEND_SMS, smsIconMap)
     {}
 
     MicrophoneIcon::MicrophoneIcon(Item *parent, std::uint32_t x, std::uint32_t y)
-        : Icon(parent, x, y, callAppStyle::icon::x_margin, MicrophoneIconState::MUTE, microphoneIconMap)
+        : StateIcon(parent, x, y, callAppStyle::icon::x_margin, MicrophoneIconState::MUTE, microphoneIconMap)
     {}
 
     SpeakerIcon::SpeakerIcon(Item *parent, std::uint32_t x, std::uint32_t y)
-        : Icon(parent, x, y, callAppStyle::icon::x_margin, SpeakerIconState::SPEAKER, speakerIconMap)
+        : StateIcon(parent, x, y, callAppStyle::icon::x_margin, SpeakerIconState::SPEAKER, speakerIconMap)
     {}
 
 } // namespace gui