+++ changelog.md
@@ -15,6 +15,8 @@
 * `[PowerManagement]` Added CPU load measurement.
 * `[alarms]` Added new/edit alarm window
 * `[file indexer]` Support for deleting entries in File Indexer DB.
+* `[listview]` Added onEmpty list callbacks and implemented them for notes and messages. 
+* `[notes]` Added list rebuild on notifications.
 
 ### Changed